Cache session insights
