+++ backend/private-graph/graph/model/models_gen.go
@@ -661,11 +661,6 @@ type SessionCommentTagInput struct {
 	Name string `json:"name"`
 }
 
-type SessionInsight struct {
-	ID      int    `json:"id"`
-	Insight string `json:"insight"`
-}
-
 type SessionQuery struct {
 	ID        int `json:"id"`
 	ProjectID int `json:"project_id"`
