+++ backend/private-graph/graph/schema.graphqls
@@ -1075,6 +1075,7 @@ type SessionComment {
 
 type SessionInsight {
 	id: ID!
+	session_id: Int!
 	insight: String!
 }
